pom/show

+++ src/main/java/hiber/service/CarService.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface CarService {
     public void add(Car car);
     public List<Car> listCar();
-    public void delete(int id);
+    public void delete(long id);
     public Car update( Car car );
     public Car findOne( long id );
